+++ example/lib/circle_annotations.dart
@@ -36,6 +36,14 @@ class CircleAnnotationPageBodyState extends State<CircleAnnotationPageBody> {
 
   _onMapCreated(MapboxMap mapboxMap) {
     this.mapboxMap = mapboxMap;
+    mapboxMap.setCamera(CameraOptions(
+      center:
+        Point(coordinates: Position(
+          0, 0
+        )).toJson(),
+      zoom: 1,
+      pitch: 0)
+    );
     mapboxMap.annotations.createCircleAnnotationManager().then((value) {
       circleAnnotationManager = value;
       createOneAnnotation();